When presenter a viewController, pip hidden fix
