--- Example/ExampleTests/ExampleTests.swift
@@ -1,34 +1,0 @@
-//
-//  ExampleTests.swift
-//  ExampleTests
-//
-//  Created by Taeun Kim on 07/12/2018.
-//  Copyright © 2018 Kofktu. All rights reserved.
-//
-
-import XCTest
-@testable import Example
-
-class ExampleTests: XCTestCase {
-
-    override func setUp() {
-        // Put setup code here. This method is called before the invocation of each test method in the class.
-    }
-
-    override func tearDown() {
-        // Put teardown code here. This method is called after the invocation of each test method in the class.
-    }
-
-    func testExample() {
-        // This is an example of a functional test case.
-        // Use XCTAssert and related functions to verify your tests produce the correct results.
-    }
-
-    func testPerformanceExample() {
-        // This is an example of a performance test case.
-        self.measure {
-            // Put the code you want to measure the time of here.
-        }
-    }
-
-}
