--- Example/Podfile
@@ -1,8 +1,0 @@
-# Uncomment the next line to define a global platform for your project
-# platform :ios, '9.0'
-use_frameworks!
-
-target 'Example' do
-  # Comment the next line if you're not using Swift and don't want to use dynamic frameworks
-  pod 'PIPKit', :path => '../'
-end
