+++ PIPKit/PIPKit.h
@@ -1,0 +1,18 @@
+//
+//  PIPKit.h
+//  PIPKit
+//
+//  Created by Kofktu on 2022/01/03.
+//
+
+#import <Foundation/Foundation.h>
+
+//! Project version number for PIPKit.
+FOUNDATION_EXPORT double PIPKitVersionNumber;
+
+//! Project version string for PIPKit.
+FOUNDATION_EXPORT const unsigned char PIPKitVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <PIPKit/PublicHeader.h>
+
+
